Moving methods from Model to _BlockData
This allows arbitrary blocks to be solved as if they were models.
